chore: refactoring & stuck notification fix
